requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
requests>=2.32.3
psycopg2-binary>=2.9.10